Tutorial scripts, images and text added. Needs some tweaking to get the order working.

+++ Assets/Script/CreatePopup.cs
@@ -5,7 +5,7 @@ using System.Collections;
 public class CreatePopup : MonoBehaviour {
 
     public string message;
-    public float duration;
+    public int duration;
     public Sprite image;
     
     void Start () {
